requirements.txt
allure-pytest==2.9.45
selene>=2.0.0rc4
pytest~=7.4.3
Appium-Python-Client~=3.1.0
python-dotenv~=1.0.0